# backend/api-doc.yaml
openapi: 3.0.0
info:
  title: Kulhad Flask Backend API
  version: 1.0.0
  description: |
    OpenAPI 3.0 specification generated from the current Flask backend blueprints.

    Authentication-protected endpoints use JWT bearer authentication.
    The computer vision endpoints upload an image to the backend, which then calls an external object-detection service internally.
servers:
  - url: http://localhost:5000
    description: Local development server
...
paths:
...
  /api/admin/users:
    get:
      tags:
        - admin
      summary: List managed users
      description: Returns customer, employee, and admin users. Supports optional role filtering.
      security:
        - bearerAuth: []
      parameters:
        - name: role
          in: query
          required: false
          description: Optional role filter.
          schema:
            type: string
            enum:
              - customer
              - employee
              - admin
      responses:
        '200':
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminUser'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags:
        - admin
      summary: Create a managed user
      description: Creates a customer or employee account. Admin access is required.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUserCreateRequest'
            example:
              name: Worker Two
              email: worker.two@example.com
              password: workerpass
              role: employee
              phone: "1234567890"
              address: Factory Road
              department: Production
              job_title: Painter
              daily_wage: 18000
              is_active: true
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUser'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/admin/users/{user_id}:
    put:
      tags:
        - admin
      summary: Update a managed user
      description: Updates an existing customer or employee account. Admin access is required.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUserUpdateRequest'
            example:
              name: Customer User Updated
              phone: "1111111111"
              address: Updated Street
              is_active: true
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUser'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/UserNotFound'
    delete:
      tags:
        - admin
      summary: Delete a managed user
      description: Deletes a user account. The current admin account cannot delete itself.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: User deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: User deleted
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/UserNotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    UserId:
      name: user_id
      in: path
      required: true
      schema:
        type: integer
      description: User identifier.
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AuthErrorResponse'
          example:
            msg: Missing Authorization Header
    Forbidden:
      description: Authenticated user does not have the required role
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageResponse'
          example:
            message: Access forbidden
    Error400:
      description: Bad request
      content:
        application/json:
          schema:
            oneOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - $ref: '#/components/schemas/MessageResponse'
    UserNotFound:
      description: User not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: User not found
...
  schemas:
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    AuthErrorResponse:
      type: object
      properties:
        msg:
          type: string
      required:
        - msg
...
    AdminUser:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum:
            - admin
            - customer
            - employee
        is_active:
          type: boolean
        created_at:
          type: string
        phone:
          type: string
        address:
          type: string
        department:
          type: string
        job_title:
          type: string
        per_unit_wage:
          type: number
          format: float
        daily_wage:
          type: number
          format: float
      required:
        - id
        - name
        - email
        - role
        - is_active
        - created_at
        - phone
        - address
        - department
        - job_title
    AdminUserCreateRequest:
      type: object
      required:
        - name
        - email
        - password
        - role
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        role:
          type: string
          enum:
            - customer
            - employee
        is_active:
          type: boolean
          default: true
        phone:
          type: string
        address:
          type: string
        department:
          type: string
        job_title:
          type: string
        per_unit_wage:
          type: number
          format: float
        daily_wage:
          type: number
          format: float
    AdminUserUpdateRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        is_active:
          type: boolean
        phone:
          type: string
        address:
          type: string
        department:
          type: string
        job_title:
          type: string
        per_unit_wage:
          type: number
          format: float
        daily_wage:
          type: number
          format: float
...